pyproject.toml
[project]
name = "ppp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nvidia-cutlass-dsl>=4.1.0",
    "torch>=2.8.0",
]
[tool.ruff]
line-length = 120